[feat] 조직도 component로 처리 후 레벨 구분선 추가 #20
